Throw away phantomjs scraper
It's much faster than loading up Firefox in XVFB, but it also means we're having
to manage two separate tools for doing the same job (load a page, including JS,
wait a few seconds and dump out the HTML).

Since phantomjs seems to be getting blocked by some hosts, we need the Firefox
scraper. Hence we might as well use it for everything that needs JS (we can
still use curl and wget if we don't care about blockers or JS).