Alter the way we read Firefox outputs
Firstly, we add a sleep to the loop. Just because we give 'select' a 0.1s
timeout doesn't mean that it takes that long to run; that's just the maximum it
will take. Hence we were seeing very high CPU loads, caused by our loop running
very quickly. Adding a sleep stops this high CPU usage.

Secondly, we were conflating two loops: reading chunks from the file descriptor
and checking if we're meant to die. Especially with the new sleep in the loop,
it's easy for the pipe to get backed up with lots of data, while we read one
chunk at a time and sleep in between. If we break the loop when told to die,
this can leave a lot of output in the pipe.

Now, we only break the loop when we've been asked to die *and* the pipe has
become empty. That way, we read out everything that was written.