Have queued only invoke a single nix-shell process
We were invoking nix-shell every 4 seconds, which is very wasteful. Instead, we
now run a loop inside a single nix-shell process.

We use a temporary directory to coordinate both loops: as long as it still
exists, the loops keep going. If the outer command is killed, its exit trap
deletes the directory, causing the loops to finish next time around.