Add captureKey
This intercepts all presses of a given key (e.g. space) until the process is
sent a TERM signal.

I wrote this to try and trap a rogue space keypress triggered by space2ctrl, but
it didn't seem to work (I think because space2ctrl sends a custom mapped keycode
255, rather than a "real" space).

Still, this command works so I'm committing it for posterity.