Add fold-unbuffered
This is a copy of 'fold.c' from GNU coreutils. It's been stripped down so that
we can avoid needing a bunch of libraries at build time (my knowledge of C,
Make, etc. is quite limited!); this is mostly just ripping out the options
system.

The major change we've made is to not buffer lines of input. Instead, we just
immediately spit out each character as we see them (possibly with "\n\r"
inserted, as needed). This allows us to get rid of the buffer completely, and
hence end up with no runtime allocations. The existing counting logic remains,
although it's simplified to e.g. remove byte-counting (we hard-code 'column
counting') and to remove 'space mode' (which would be nice, but involves
backtracking, which we can't do now that we're immediately outputting chars).