Have xvfb-run-safe pick display number at random
I've been hitting a problem that might be a race condition, where the error
messages are like:

    _XSERVTransSocketUNIXCreateListener: ...SocketCreateListener() failed
    _XSERVTransMakeAllCOTSServerListeners: server already running

I *think* this is due to one xvfb command not being quite finished yet before
relinquishing its resources.

I can't pin down an exact fix, but one way to make it less common is to
randomise the order that we try server numbers; at the moment we're almost
guaranteed for processes to clobber each other.