markRead: Pull real email addresses out of 'vanity addresses'
The markRead command takes a filename as argument, which is very specific, but
we can't use it in our mu queries. Instead, we take an indirect route via the
'From:' header: we read the 'From:' address from that file, then query the
entire mu database for the ID and filename of all entries with that
from-address. We filter this down to the filename we have, then mark that ID as
being read.

Unfortunately this might go wrong if the 'From:' header contains "vanity
addresses", like '"Chris Warburton" <chriswarbo@example.com>', since these won't
match if we query mu for them.

This change checks to see if the extracted address is of this form, and if so we
take only the part between the '<' and '>'. This seems to work, so it's OK for
now.