Splice 'fail' into builder, rather than passing to attrsToDirs
bin/fail is a symlink to the wrapped fail script. Adding "${self.fail}/bin/fail"
to an attrsToDirs call seems to produce the symlink but not the original file;
hence we get an error when trying to dereference it.

Longer term we should add string context to attrsToDirs, but for now we can just
splice the path into our builder.