Make keepassxc-proxy-alternative threaded
If KeePassXC sends something to the socket, we should read it and send to
the browser plugin over stdout.

If the browser plugin sends something via stdin, we should read it and send to
KeePassXC via the socket.

These messages aren't necessarily in lock-step (i.e. request/response pairs), so
we should always be checking both input sources (stdin and the socket).

I've added threading so we can check both concurrently.